Move `filesystem` up a dir level.
