+++ packages/breadboard/src/file-system/assets.ts
