+++ packages/breadboard/src/file-system/blob-transform.ts
