+++ packages/breadboard/src/file-system/ephemeral-blob-store.ts
