+++ packages/breadboard/src/file-system/stub-file-system.ts
